third_party/WebKit/LayoutTests/hittesting/inline-with-clip-path-expected.txt

Inline hit testing should not include clipped-out regions.

 a a
PASS successfullyParsed is true

TEST COMPLETE
PASS document.elementFromPoint(referenceClipPathRect.left + 25, referenceClipPathRect.top + 75).id is not 'referenceClipPath'
PASS document.elementFromPoint(referenceClipPathRect.left + 75, referenceClipPathRect.top + 25).id is 'referenceClipPath'
PASS document.elementFromPoint(shapeClipPathRect.left + 25, shapeClipPathRect.top + 75).id is not 'shapeClipPath'
PASS document.elementFromPoint(shapeClipPathRect.left + 75, shapeClipPathRect.top + 25).id is 'shapeClipPath'
